Q1: move PlaylistEntry to neutral model owner
Preserve the 70-line class byte-for-byte and all raw constructor/copy quirks pinned by 8933bca8. Retain the legacy path as a same-type export and change only fifteen consumer imports. The separate identity test proves typed-list interchange. Same-checker layering falls 77 to 62 with zero additions; the final four-file suite passes 24 cases. Origin pin bytes remain unchanged.
