+++ lib/services/cloud/alldebrid_cloud_provider.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/foundation.dart';
 
 import '../../models/torrent.dart';
-import '../../screens/video_player/models/playlist_entry.dart';
+import '../../models/playlist_entry.dart';
 import '../../utils/file_utils.dart';
 import '../../utils/stremio_episode_selector.dart';
 import '../alldebrid_service.dart';
