+++ lib/services/cloud/cloud_capabilities.dart
@@ -1,5 +1,5 @@
 import '../../models/torrent.dart';
-import '../../screens/video_player/models/playlist_entry.dart';
+import '../../models/playlist_entry.dart';
 import '../series_source_service.dart';
 import 'cloud_playback_result.dart';
 import 'cloud_playlist_result.dart';
