+++ lib/services/cloud/pikpak_cloud_provider.dart
@@ -3,7 +3,7 @@ import 'dart:math';
 import 'package:flutter/foundation.dart';
 
 import '../../models/torrent.dart';
-import '../../screens/video_player/models/playlist_entry.dart';
+import '../../models/playlist_entry.dart';
 import '../../utils/file_utils.dart';
 import '../../utils/series_parser.dart';
 import '../../utils/stremio_episode_selector.dart';
