+++ lib/widgets/movie_collection_browser.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import '../models/movie_collection.dart';
-import '../screens/video_player/models/playlist_entry.dart';
+import '../models/playlist_entry.dart';
 import '../services/storage_service.dart';
 
 class MovieCollectionBrowser extends StatefulWidget {
